requirements.cpu.txt
# python==3.10
requests~=2.31.0
fastapi~=0.100.0
uvicorn~=0.22.0
openai~=0.27.8
sse-starlette~=1.6.1
pydantic~=1.10.7
transformers_stream_generator~=0.0.4